backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.3
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
dnspython==2.5.0
drf-spectacular==0.27.0
eventlet==0.35.1
flower==2.0.1
greenlet==3.0.3
humanize==4.9.0
inflection==0.5.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kombu==5.3.5
model-mommy==2.0.0
pillow==10.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
pycodestyle==2.10.0
pydocstyle==6.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
